Implementing grass runup Rayleigh. DIKERNEL-69

+++ DiKErnel/Src/DiKErnel.FunctionLibrary/Public/GrassRevetmentWaveRunupRayleigh.h
@@ -58,7 +58,7 @@ namespace DiKErnel::FunctionLibrary
                 double criticalFrontVelocity,
                 double increasedLoadTransitionAlphaM,
                 double reducedStrengthTransitionAlphaS,
-                double cumulativeOverloadNf
+                int cumulativeOverloadNf
             );
     };
 }
